Now all pages go straight to auth first, ws using env value, and improved structure.
