+++ src/App.css
@@ -79,16 +79,3 @@ footer {
   bottom: 0;
 }
 
-/*@media screen and (max-width: 768px) {*/
-/*  #root {*/
-/*    max-width: 768px;*/
-/*    margin: 0 auto;*/
-/*    padding: 0;*/
-/*    text-align: center;*/
-/*  }*/
-
-/*  .card {*/
-/*    padding: 0;*/
-/*  }*/
-/*}*/
-
