+++ src/components/ProtectedComponent.tsx
@@ -1,0 +1,13 @@
+import {FC} from "react";
+import {withAuthenticationRequired} from "@auth0/auth0-react";
+
+interface ProtectedComponentProps {
+    component: FC;
+    propsForComponent?: any
+}
+const ProtectedComponent: FC<ProtectedComponentProps> = ({component, propsForComponent}) => {
+    const Protected =withAuthenticationRequired(component)
+    return <Protected {...propsForComponent} />
+}
+
+export default ProtectedComponent
